requirements.txt
yfinance>=0.2.36
pandas>=2.2.0
SQLAlchemy>=2.0.27
python-dotenv>=1.0.1
pytest>=8.0.0
loguru>=0.7.2
numpy>=1.24.0
pyarrow>=15.0.0
fastparquet>=2024.2.0
scikit-learn>=1.4.0
lightgbm>=4.3.0
optuna>=3.5.0
joblib>=1.3.2 